Extract `ImagePreviewView` + add `#Preview`
